@@ -23,7 +23,8 @@
     "@dnd-kit/sortable": "^10.0.0",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
-    "@loadable/component": "^5.16.4",
+    "@eslint/js": "^9.28.0",
+    "@loadable/component": "^5.16.7",
     "@mui/icons-material": "^7.0.2",
     "@mui/material": "^7.0.2",
     "@mui/x-date-pickers": "^7.28.3",
@@ -35,19 +36,18 @@
     "@pdg/react-hook": "^1.0.28",
     "@pdg/react-table": "^1.0.131",
     "@pdg/util": "^1.0.23",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.5.16",
+    "@pmmmwh/react-refresh-webpack-plugin": "^0.6.0",
     "@soda/friendly-errors-webpack-plugin": "^1.8.1",
     "@tinymce/tinymce-react": "6.1.0",
     "@types/jest": "^29.5.14",
     "@types/loadable__component": "^5.13.9",
-    "@types/node": "^22.14.1",
-    "@types/react": "^19.1.1",
-    "@types/react-dom": "^19.1.2",
+    "@types/node": "^22.15.29",
+    "@types/react": "^19.1.6",
+    "@types/react-dom": "^19.1.5",
     "@types/react-router-dom": "^5.3.3",
     "@types/uuid": "^10.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.29.1",
-    "@typescript-eslint/parser": "^8.29.1",
-    "axios": "^1.8.4",
+    "@typescript-eslint/parser": "^8.33.0",
+    "axios": "^1.9.0",
     "classnames": "^2.5.1",
     "clean-webpack-plugin": "^4.0.0",
     "copy-to-clipboard": "^3.3.3",
@@ -56,13 +56,13 @@
     "css-loader": "^7.1.2",
     "dayjs": "^1.11.13",
     "dotenv": "^16.5.0",
-    "eslint": "8.57.1",
-    "eslint-config-prettier": "^10.1.2",
+    "eslint": "9.28.0",
+    "eslint-config-prettier": "^10.1.5",
     "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-prettier": "^5.2.6",
+    "eslint-plugin-prettier": "^5.4.1",
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^5.2.0",
-    "eslint-webpack-plugin": "^4.2.0",
+    "eslint-webpack-plugin": "^5.0.1",
     "fork-ts-checker-webpack-plugin": "^9.1.0",
     "html-loader": "^5.1.0",
     "html-webpack-plugin": "^5.6.3",
@@ -74,28 +74,24 @@
     "react-dom": "^19.1.0",
     "react-number-format": "^5.4.4",
     "react-resize-detector": "^12.0.2",
-    "react-router": "^7.5.0",
-    "react-router-dom": "^7.5.0",
+    "react-router": "^7.6.1",
+    "react-router-dom": "^7.6.1",
     "react-use-error-boundary": "^3.0.0",
-    "sass": "^1.86.3",
+    "sass": "^1.89.1",
     "sass-loader": "^16.0.5",
     "simplebar-react": "^3.3.0",
     "source-map-loader": "^5.0.0",
     "style-loader": "^4.0.0",
-    "styled-components": "^6.1.17",
+    "styled-components": "^6.1.18",
     "ts-loader": "^9.5.2",
     "typescript": "5.8.3",
+    "typescript-eslint": "^8.33.0",
     "url-loader": "^4.1.1",
     "usehooks-ts": "^3.1.1",
     "uuid": "^11.1.0",
-    "webpack": "^5.99.5",
+    "webpack": "^5.99.9",
     "webpack-bundle-analyzer": "^4.10.2",
     "webpack-cli": "^6.0.1",
     "webpack-dev-server": "^5.2.1"
-  },
-  "overrides": {
-    "@loadable/component": {
-      "react": "^19.0.0"
-    }
   }
 }
