@@ -37,7 +37,7 @@
   "eslintConfig": {},
   "devDependencies": {
     "cross-env": "^7.0.3",
-    "eslint": "8.57.1",
+    "eslint": "9.28.0",
     "prettier": "^3.5.3"
   }
 }
