@@ -33,44 +33,44 @@
     "express": "^4.21.2"
   },
   "devDependencies": {
+    "@eslint/js": "^9.28.0",
     "@pdg/util": "^1.0.23",
     "@types/cookie-parser": "^1.4.8",
     "@types/csurf": "^1.11.5",
     "@types/ejs": "^3.1.5",
-    "@types/express": "^4.17.21",
+    "@types/express": "^4.17.22",
     "@types/express-session": "^1.18.1",
-    "@types/node": "^20.12.11",
+    "@types/node": "^22.15.29",
     "@types/webpack-env": "^1.18.8",
-    "@typescript-eslint/eslint-plugin": "^8.29.1",
-    "@typescript-eslint/parser": "^8.29.1",
+    "@typescript-eslint/parser": "^8.33.0",
     "agentkeepalive": "^4.6.0",
-    "axios": "^1.8.4",
+    "axios": "^1.9.0",
     "clean-webpack-plugin": "^4.0.0",
     "concurrently": "^9.1.2",
     "connect-redis": "^7.1.1",
     "cookie-parser": "^1.4.7",
-    "core-js": "^3.41.0",
+    "core-js": "^3.42.0",
     "cross-env": "^7.0.3",
     "csurf": "^1.11.0",
     "dotenv": "^16.5.0",
     "ejs": "^3.1.10",
-    "eslint": "8.57.1",
-    "eslint-config-prettier": "^10.1.2",
-    "eslint-plugin-jsx-a11y": "^6.10.2",
-    "eslint-plugin-prettier": "^5.2.6",
+    "eslint": "9.28.0",
+    "eslint-config-prettier": "^10.1.5",
+    "eslint-plugin-prettier": "^5.4.1",
     "express-session": "^1.18.1",
     "fork-ts-checker-webpack-plugin": "^9.1.0",
-    "helmet": "^7.2.0",
+    "helmet": "^8.1.0",
     "http-proxy-middleware": "^3.0.5",
     "nocache": "^4.0.0",
-    "node-hot-loader": "^1.21.11",
-    "nodemon": "^3.1.9",
+    "node-hot-loader": "^1.23.0",
+    "nodemon": "^3.1.10",
     "path": "^0.12.7",
     "prettier": "^3.5.3",
-    "redis": "^4.7.0",
+    "redis": "^5.1.1",
     "ts-loader": "^9.5.2",
     "typescript": "5.8.3",
-    "webpack": "^5.99.5",
+    "typescript-eslint": "^8.33.0",
+    "webpack": "^5.99.9",
     "webpack-cli": "^6.0.1"
   }
 }
