@@ -1,12 +1,12 @@
-module.exports = {
+export default {
   printWidth: 120,
   tabWidth: 2,
   useTabs: false,
   semi: true,
   singleQuote: true,
   jsxSingleQuote: true,
-  trailingComma: "es5",
+  trailingComma: 'es5',
   bracketSpacing: true,
   jsxBracketSameLine: false,
-  arrowParens: "always",
+  arrowParens: 'always',
 };
